+++ src/main/java/xyz/superez/dynamickeepinv/IntegrationManager.java
@@ -6,6 +6,8 @@ import xyz.superez.dynamickeepinv.hooks.GravesXHook;
 import xyz.superez.dynamickeepinv.hooks.GriefPreventionHook;
 import xyz.superez.dynamickeepinv.hooks.LandsHook;
 import xyz.superez.dynamickeepinv.hooks.MMOItemsHook;
+import xyz.superez.dynamickeepinv.hooks.TownyHook;
+import xyz.superez.dynamickeepinv.hooks.WorldGuardHook;
 
 public class IntegrationManager {
 
@@ -13,6 +15,8 @@ public class IntegrationManager {
 
     private LandsHook landsHook;
     private GriefPreventionHook griefPreventionHook;
+    private WorldGuardHook worldGuardHook;
+    private TownyHook townyHook;
     private GravesXHook gravesXHook;
     private AxGravesHook axGravesHook;
     private MMOItemsHook mmoItemsHook;
@@ -32,6 +36,18 @@ public class IntegrationManager {
             griefPreventionHook = new GriefPreventionHook(plugin);
         } else {
             griefPreventionHook = null;
+        }
+
+        if (plugin.getConfig().getBoolean("integrations.worldguard.enabled", false)) {
+            worldGuardHook = new WorldGuardHook(plugin);
+        } else {
+            worldGuardHook = null;
+        }
+
+        if (plugin.getConfig().getBoolean("integrations.towny.enabled", false)) {
+            townyHook = new TownyHook(plugin);
+        } else {
+            townyHook = null;
         }
 
         if (plugin.getConfig().getBoolean("integrations.gravesx.enabled", false)) {
@@ -84,6 +100,22 @@ public class IntegrationManager {
 
     public boolean isGriefPreventionEnabled() {
         return griefPreventionHook != null && griefPreventionHook.isAvailable();
+    }
+
+    public WorldGuardHook getWorldGuardHook() {
+        return worldGuardHook;
+    }
+
+    public boolean isWorldGuardEnabled() {
+        return worldGuardHook != null && worldGuardHook.isAvailable();
+    }
+
+    public TownyHook getTownyHook() {
+        return townyHook;
+    }
+
+    public boolean isTownyEnabled() {
+        return townyHook != null && townyHook.isAvailable();
     }
 
     public GravesXHook getGravesXHook() {